pyproject.toml
[project]
name = "fbmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
